Implement localStorage and refactor date range filtering
